PWA-2540 Moved coupon styles to dedicated file.
